fix submit problem

+++ lib/cuckoo/common/demux.py
@@ -203,7 +203,7 @@ def demux_sflock(filename, options, package):
     except Exception as e:
         log.error(e, exc_info=True)
 
-    if cuckoo_conf.cuckoo.delete_archive:
+    if cuckoo_conf.cuckoo.delete_archive and retlist:
         os.remove(filename)
 
     return list(filter(None, retlist))
